Yeah

+++ Events/Filter.js
@@ -5,8 +5,9 @@ const config = require('../Data/config');
 const admin = require('firebase-admin');
 //const leet = require('l33tsp34k');
 const DataBase = admin.firestore();
+const { v4: uuidv4 } = require('uuid');
 
-function CleanMessage(Content) {
+function clean(Content) {
     if (typeof Content === 'string') {
         Content = Content
             .replace(/`/g, `\`${String.fromCharCode(8203)}`)
@@ -26,8 +27,10 @@ function strip(Content) {
     return Content;
 }
 
+const CharSet = ['a','b','c','d','e','f','g','h','i','j','k','l','m','n','o','p','q','r','s','t','u','v','w','x','y','z','~','!','@','#','$','%','^','&','*','(',')','_','+','`','1','2','3','4','5','6','7','8','9','0','-','=','[',']','{','}',';','\'','\\',':','"','|',',','.','/','<','>','?', ' ', '\n'];
 
-module.exports ={
+
+module.exports = {
     Start: async (client) => {
         let RawData = await DataBase.collection('config').doc('filter').get();
         let Data = RawData.data();
@@ -35,18 +38,43 @@ module.exports ={
         async function filter(message) {
             if (message.author.bot || !message.guild || message.system) return;
 
+            // let Index_2 = message.content.toLowerCase();
+
+            // for (var j = 0; j < Index_2.length; j++) {
+            //     let char = Index_2.charAt(j);
+
+            //     if (!CharSet.includes(char)) {
+            //         Index_2.slice(j);
+            //     }
+
+            //     //print(Index_2);
+            // }
+
             if (message.member.roles.cache.has(config.bypass)) return;
+
+            let Index = message.content.toLowerCase();
+
+
+            for (var i = 0; i < Index.length; i++) {
+                let char = Index.charAt(i);
+
+                if (!CharSet.includes(char)) message.delete();
+            }
+
+
+
             let Message = strip(message.content.toLowerCase().replaceAll(' ', ''));
             //let JoinedMessage = Message.replaceAll(' ', '');
 
             // Blocked Words
             for (var word of Data.blocked_words) {
                 let Matched = Message.match(new RegExp(word));
+
                 if (Matched) {
 
                     if (Data.allowed_words.includes(Message)) return print(`Word Included`);
 
-                    message.delete();
+                    message.delete().catch();
 
                     //let Striked = await strike(message.author);
 
@@ -54,7 +82,7 @@ module.exports ={
                         title: `Please don't do that!`,
                         description: `We try to keep the server clean and it would be nice if you didn't send stuff like that. Thanks!`,//\n\nYou now have ${Striked} strikes.`,
                         color: `RED`
-                    })).then(m=>{m.delete({timeout:5000});});
+                    })).then(m=>{m.delete({timeout:5000}).catch();});
 
                     return client.channels.cache.get(config.logs).send({
                         embed: {
@@ -67,9 +95,9 @@ module.exports ={
                                 Channel: ${message.channel}
                                 Message ID: ${message.id}
 
-                                Original Message: \`\`\`${CleanMessage(message.content)}\`\`\`
-                                String: \`\`\`${CleanMessage(Message)}\`\`\`
-                                Matched Word: \`\`\`${CleanMessage(word)}\`\`\`
+                                Original Message: \`\`\`${clean(message.content)}\`\`\`
+                                Method: \`String Detection\`: \`\`\`${clean(Message)}\`\`\`
+                                Matched Word: \`\`\`${clean(word)}\`\`\`
                             `,
                             color: `RED`
                         }
@@ -83,7 +111,7 @@ module.exports ={
                 let Matched = Message.match(new RegExp(_word));
 
                 if (Matched) {
-                    message.delete();
+                    message.delete().catch();
 
                     client.channels.cache.get(config.logs).send({
                         embed: {
@@ -96,9 +124,9 @@ module.exports ={
                                 Channel: ${message.channel}
                                 Message ID: ${message.id}
 
-                                Original Message: \`\`\`${CleanMessage(message.content)}\`\`\`
-                                String: \`\`\`${CleanMessage(Message)}\`\`\`
-                                Matched Word: \`\`\`${CleanMessage(_word)}\`\`\`
+                                Original Message: \`\`\`${clean(message.content)}\`\`\`
+                                Method: String Detection: \`\`\`${clean(Message)}\`\`\`
+                                Matched Word: \`\`\`${clean(_word)}\`\`\`
                             `,
                             color: `RED`
                         }
@@ -123,7 +151,7 @@ module.exports ={
 
 
         client.on('message', async (message) => {
-            if(message.member.roles.cache.has(`777258405675925524`)) return message.delete();
+            if(message.member.roles.cache.has(`777258405675925524`)) message.delete();
 
             filter(message);
         });
